% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compopsizeVN.R
\docType{package}
\name{compopsizeVN}
\alias{compopsizeVN}
\alias{compopsizeVN-package}
\title{compopsizeVN: Polygons and populations sizes of Vietnamese communes.}
\description{
The compopsizeVN package provides polygons and populations sizes of the communes of Vietnam.
}

